refactor: moved User DTOs
