+++ src/main/java/com/booklink/backend/dto/user/CreateUserDto.java
@@ -1,4 +1,4 @@
-package com.booklink.backend.dto;
+package com.booklink.backend.dto.user;
 
 import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
